requirements.txt
# cycler==0.10.0
# future==0.18.2
# iso8601==0.1.16
# kiwisolver==1.3.1
# matplotlib==3.4.3
# numpy==1.21.2
# pandas==1.3.2
# picosdk==1.0
# Pillow==8.3.1
# pyparsing==2.4.7
# pyserial==3.5
# python-dateutil==2.8.2
# pytz==2021.1
# PyYAML==5.4.1
# scipy==1.7.1
# six==1.16.0

future==0.18.2
iso8601==0.1.16
matplotlib==3.4.3
pandas==1.3.2
picosdk==1.0
pyserial==3.5
pyyaml==5.4.1
scipy==1.7.1

numpy~=1.21.2
requests~=2.26.0
websockets~=10.0